// frontend/src/app/api/generated/model/admission-dto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AdmissionDto { 
    uid?: string;
    patientUid?: string;
    wardBedUid?: string;
    paymentType?: string;
    insurancePlanUid?: string;
    membershipNo?: string;
    status?: string;
    admittedAt?: string;
    dischargedAt?: string;
}